+++ src/web/005_request.js
@@ -12,7 +12,21 @@ function Request(options) {
 	// If there are any fully-uppercase keys, treat it like the method/url pair
 	if (!options.method && !options.url) {
 		for (var k in options) {
-			if (k.toUpperCase() == k) {
+			// 'METHOD url'
+			if (k.indexOf(' ') !== -1) {
+				var kparts = k.split(' ');
+				if (kparts.length !== 2) {
+					console.warn('Invalid request key:', k, this);
+					continue;
+				}
+				options.method = kparts[0];
+				options.url = kparts[1];
+				options.body = options[k];
+				delete options[k];
+				break;
+			}
+			// 'METHOD'
+			else if (k.toUpperCase() == k) {
 				options.method = k;
 				options.url = options[k];
 				delete options[k];
@@ -31,7 +45,6 @@ function Request(options) {
 	this.host = options.host || null;
 	this.query = options.query || {};
 	this.headers = lowercaseKeys(headers);
-	this.body = '';
 
 	// Guess the content-type if a full body is included in the message
 	if (options.body && !this.headers['content-type']) {
@@ -50,7 +63,7 @@ function Request(options) {
 		writable: true
 	});
 	Object.defineProperty(this, 'body', {
-		value: '',
+		value: options.body || '',
 		configurable: true,
 		enumerable: false,
 		writable: true