+++ test/web/core.js
@@ -222,6 +222,27 @@ success
 }
 */
 
+// 'METHOD url': body
+
+done = false;
+startTime = Date.now();
+var res = local.dispatch({
+  'POST httpl://test.com/foo': 'echo this, please',
+  Accept: 'text/plain', Content_Type: 'text/plain'
+});
+res.then(printSuccess, printError).always(finishTest);
+wait(function () { return done; });
+
+/* =>
+success
+{
+  body: "echo this, please",
+  headers: {"content-type": "text/plain"},
+  reason: "ok",
+  status: 200
+}
+*/
+
 // uppercase headers in responses
 
 done = false;